int32(INT_FAST16_MAX)

+++ src/AllQuartets.cpp
@@ -31,7 +31,7 @@ IntegerMatrix all_quartets(IntegerVector nTips) {
   if (!nTips.length()) {
     Rcpp::stop("nTips must contain a single integer value");
   }
-  if (nTips[0] >= INT_FAST16_MAX) {
+  if (nTips[0] >= int32(INT_FAST16_MAX)) {
     Rcpp::stop("nTips must be < 32768");
   }
   if (nTips[0] < 4) {
